test: separated construction from other tests and removed integration for other rosetta tests
