+++ src/tests-rosetta/api.ts
@@ -1,134 +1,43 @@
 import { PgDataStore, cycleMigrations, runMigrations } from '../datastore/postgres-store';
 import { PoolClient } from 'pg';
-import * as BigNum from 'bn.js';
 import { ApiServer, startApiServer } from '../api/init';
 import * as supertest from 'supertest';
-import { startEventServer } from '../event-stream/event-server';
-import { Server } from 'net';
-import { DbBlock, DbTx, DbMempoolTx, DbTxStatus, DbTxTypeId, DbTxAnchorMode } from '../datastore/common';
+import { DbMempoolTx, DbTxStatus, DbTxTypeId, DbTxAnchorMode } from '../datastore/common';
 import * as assert from 'assert';
 import {
   AnchorMode,
-  AuthType,
-  bufferCV,
   ChainID,
-  createStacksPrivateKey,
-  getPublicKey,
-  makeSTXTokenTransfer,
-  makeUnsignedContractCall,
-  makeUnsignedSTXTokenTransfer,
-  noneCV,
-  pubKeyfromPrivKey,
-  publicKeyToString,
-  SignedTokenTransferOptions,
-  someCV,
-  standardPrincipalCV,
-  TransactionSigner,
-  tupleCV,
-  uintCV,
-  UnsignedContractCallOptions,
-  UnsignedTokenTransferOptions,
 } from '@stacks/transactions';
-import * as BN from 'bn.js';
-import { StacksCoreRpcClient } from '../core-rpc/client';
-import { bufferToHexPrefixString, getBlockTxUpdateData, timeout } from '../helpers';
 import {
-  RosettaConstructionCombineRequest,
-  RosettaConstructionCombineResponse,
   RosettaAccount,
   RosettaAccountBalanceRequest,
   RosettaAccountBalanceResponse,
-  RosettaAccountIdentifier,
   RosettaAmount,
-  RosettaConstructionDeriveRequest,
-  RosettaConstructionDeriveResponse,
-  RosettaConstructionHashRequest,
-  RosettaConstructionHashResponse,
-  RosettaConstructionMetadataRequest,
-  RosettaConstructionParseRequest,
-  RosettaConstructionParseResponse,
-  RosettaConstructionPayloadsRequest,
-  RosettaConstructionPreprocessRequest,
-  RosettaConstructionPreprocessResponse,
   RosettaMempoolRequest,
   RosettaMempoolResponse,
   RosettaMempoolTransactionRequest,
   RosettaMempoolTransactionResponse,
   RosettaOperation,
   RosettaTransaction,
-  RosettaConstructionMetadataResponse,
 } from '@stacks/stacks-blockchain-api-types';
 import {
-  getRosettaNetworkName,
-  RosettaConstants,
   RosettaErrors,
-  RosettaErrorsTypes,
   RosettaOperationTypes,
   RosettaOperationStatuses,
 } from '../api/rosetta-constants';
-import { getStacksTestnetNetwork, testnetKeys } from '../api/routes/debug';
-import {
-  getSignature,
-  getStacksNetwork,
-  makePresignHash,
-} from '../rosetta-helpers';
-import { makeSigHashPreSign, MessageSignature } from '@stacks/transactions';
-import { decodeBtcAddress } from '@stacks/stacking';
 import { TestBlockBuilder } from '../test-utils/test-builders';
 
 
 describe('Rosetta API', () => {
   let db: PgDataStore;
   let client: PoolClient;
-  let eventServer: Server;
   let api: ApiServer;
 
-  function standByForTx(expectedTxId: string): Promise<DbTx> {
-    const broadcastTx = new Promise<DbTx>(resolve => {
-      const listener: (txId: string) => void = async txId => {
-        const dbTxQuery = await api.datastore.getTx({ txId: txId, includeUnanchored: true });
-        if (!dbTxQuery.found) {
-          return;
-        }
-        const dbTx = dbTxQuery.result as DbTx;
-        if (
-          dbTx.tx_id === expectedTxId &&
-          (dbTx.status === DbTxStatus.Success ||
-            dbTx.status === DbTxStatus.AbortByResponse ||
-            dbTx.status === DbTxStatus.AbortByPostCondition)
-        ) {
-          api.datastore.removeListener('txUpdate', listener);
-          resolve(dbTx);
-        }
-      };
-      api.datastore.addListener('txUpdate', listener);
-    });
-
-    return broadcastTx;
-  }
-
-  async function standByForPoxToBeReady(): Promise<void>{
-    let tries = 0;
-    while(true){
-      try{
-        tries++;
-        await  new StacksCoreRpcClient().getPox();
-        return Promise.resolve();
-      }
-      catch(error){
-        console.log('Error getting pox info on try ' + tries, error);
-        timeout(100);
-      }
-    }
-  }
-
-  beforeAll(async () => {
+  beforeEach(async () => {
     process.env.PG_DATABASE = 'postgres';
-    process.env.STACKS_CHAIN_ID = '0x80000000';
     await cycleMigrations();
     db = await PgDataStore.connect({ usageName: 'tests' });
     client = await db.pool.connect();
-    eventServer = await startEventServer({ datastore: db, chainId: ChainID.Testnet });
     api = await startApiServer({ datastore: db, chainId: ChainID.Testnet });
   });
 
@@ -204,32 +113,35 @@ describe('Rosetta API', () => {
 
   test('network/status', async () => {
     // skip first a block (so we are at least N+1 blocks)
-    await new Promise<string>(resolve =>
-      api.datastore.once('blockUpdate', block => resolve(block))
-    );
-    const blockHash = await new Promise<string>(resolve =>
-      api.datastore.once('blockUpdate', block => resolve(block))
-    );
-    const genesisBlock = await api.datastore.getBlock({ height: 1 });
-    assert(genesisBlock.found);
+    const genesisData = new TestBlockBuilder().build();
+    const blockBuilderData = {
+      parent_block_hash: genesisData.block.block_hash,
+      parent_index_block_hash: genesisData.block.index_block_hash,
+      block_height: 2,
+      index_block_hash: '0x12345678'
+    } 
+    const blockData = new TestBlockBuilder(blockBuilderData).build();
+
+    await db.update(genesisData);
+    await db.update(blockData);
+
+    const block = blockData.block, genesisBlock = genesisData.block;
+
     const query1 = await supertest(api.address)
       .post(`/rosetta/v1/network/status`)
       .send({ network_identifier: { blockchain: 'stacks', network: 'testnet' } });
     expect(query1.status).toBe(200);
     expect(query1.type).toBe('application/json');
-    const blockQuery = await api.datastore.getBlock({ hash: blockHash });
-    assert(blockQuery.found);
-    const block = blockQuery.result;
 
     const expectResponse = {
       current_block_identifier: {
         index: block.block_height,
-        hash: blockHash,
+        hash: block.block_hash,
       },
       current_block_timestamp: block.burn_block_time * 1000,
       genesis_block_identifier: {
-        index: genesisBlock.result.block_height,
-        hash: genesisBlock.result.block_hash,
+        index: genesisBlock.block_height,
+        hash: genesisBlock.block_hash,
       },
       peers: [],
     };
@@ -248,10 +160,23 @@ describe('Rosetta API', () => {
 
   test('block - by index', async () => {
     const blockHeight = 2;
-    const block = await api.datastore.getBlock({ height: blockHeight });
-    assert(block.found);
-    const txs = await api.datastore.getBlockTxsRows(block.result.block_hash);
-    assert(txs.found);
+    const tx = {
+      type_id: DbTxTypeId.Coinbase,
+      anchor_mode: AnchorMode.OnChainOnly,
+      status: DbTxStatus.Success,
+      sender_address: 'ST2N3HQ5BZV43J0ZZY4X6T50NTQTQBMZ3AVRQ86A',
+      coinbase_payload: Buffer.alloc(0)
+    }
+    const parentData = new TestBlockBuilder().addTx().build();
+    const data = new TestBlockBuilder({
+      block_height: 2, 
+      parent_block_hash: parentData.block.block_hash, 
+      block_hash: '0x1234', 
+      parent_index_block_hash: parentData.block.index_block_hash, 
+      index_block_hash: '0x1234'
+    }).addTx(tx).build();
+    await db.update(parentData)
+    await db.update(data);
     const query1 = await supertest(api.address)
       .post(`/rosetta/v1/block`)
       .send({
@@ -264,24 +189,24 @@ describe('Rosetta API', () => {
       block: {
         block_identifier: {
           index: blockHeight,
-          hash: block.result.block_hash,
+          hash: data.block.block_hash,
         },
         parent_block_identifier: {
           index: blockHeight - 1,
-          hash: block.result.parent_block_hash,
+          hash: data.block.parent_block_hash,
         },
-        timestamp: block.result.burn_block_time * 1000,
+        timestamp: data.block.burn_block_time * 1000,
         transactions: [
           {
             transaction_identifier: {
-              hash: txs.result[0].tx_id,
+              hash: data.txs[0].tx.tx_id,
             },
             operations: [
               {
                 operation_identifier: { index: 0 },
                 type: 'coinbase',
                 status: 'success',
-                account: { address: txs.result[0].sender_address },
+                account: { address: data.txs[0].tx.sender_address },
               },
             ],
           },
@@ -292,15 +217,28 @@ describe('Rosetta API', () => {
 
   test('block - by hash', async () => {
     const blockHeight = 2;
-    const block = await api.datastore.getBlock({ height: blockHeight });
-    assert(block.found);
-    const txs = await api.datastore.getBlockTxsRows(block.result.block_hash);
-    assert(txs.found);
+    const tx = {
+      type_id: DbTxTypeId.Coinbase,
+      anchor_mode: AnchorMode.OnChainOnly,
+      status: DbTxStatus.Success,
+      sender_address: 'ST2N3HQ5BZV43J0ZZY4X6T50NTQTQBMZ3AVRQ86A',
+      coinbase_payload: Buffer.alloc(0)
+    }
+    const parentData = new TestBlockBuilder().addTx().build();
+    const data = new TestBlockBuilder({
+      block_height: 2, 
+      parent_block_hash: parentData.block.block_hash, 
+      block_hash: '0x1234', 
+      parent_index_block_hash: parentData.block.index_block_hash, 
+      index_block_hash: '0x1234'
+    }).addTx(tx).build();
+    await db.update(parentData)
+    await db.update(data);
     const query1 = await supertest(api.address)
       .post(`/rosetta/v1/block`)
       .send({
         network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { hash: block.result.block_hash },
+        block_identifier: { hash: data.block.block_hash },
       });
     expect(query1.status).toBe(200);
     expect(query1.type).toBe('application/json');
@@ -308,24 +246,24 @@ describe('Rosetta API', () => {
       block: {
         block_identifier: {
           index: blockHeight,
-          hash: block.result.block_hash,
+          hash: data.block.block_hash,
         },
         parent_block_identifier: {
           index: blockHeight - 1,
-          hash: block.result.parent_block_hash,
+          hash: data.block.parent_block_hash,
         },
-        timestamp: block.result.burn_block_time * 1000,
+        timestamp: data.block.burn_block_time * 1000,
         transactions: [
           {
             transaction_identifier: {
-              hash: txs.result[0].tx_id,
+              hash: data.txs[0].tx.tx_id,
             },
             operations: [
               {
                 operation_identifier: { index: 0 },
                 type: 'coinbase',
                 status: 'success',
-                account: { address: txs.result[0].sender_address },
+                account: { address: data.txs[0].tx.sender_address },
               },
             ],
           },
@@ -335,10 +273,26 @@ describe('Rosetta API', () => {
   });
 
   test('block - get latest', async () => {
-    const block = await api.datastore.getCurrentBlock();
-    assert(block.found);
-    const txs = await api.datastore.getBlockTxsRows(block.result.block_hash);
-    assert(txs.found);
+    const block1 = {
+      block_hash: '0xf989ae0dd9df72ba3256737cc088579fda6c9ad3779515e1f88f6f9b6d68a4af',
+      index_block_hash: '0x2d1a360199f2992176b6c517df02e0350de863ac1d704e27f873aca41d1d5c58',
+      burn_block_time: 1647245793,
+      burn_block_hash: '0x12771355e46cd47c71ed1721fd5319b383cca3a1f9fce3aa1c8cd3bd37af20d7',
+    }
+
+    const block2 = {
+      block_hash: '0x152ea79ffe71b835ccb575b732ed70354b169d44c6317752b2e1d48d15decf83',
+      index_block_hash: '0x0d5247fb314341a505fe27740d7bd84fe3f071ff6b06d3ccfb8a00ba4d51b84f',
+      parent_index_block_hash: block1.index_block_hash,
+      parent_block_hash: block1.block_hash,
+      block_height: 2,
+      burn_block_time: 94869286,
+      burn_block_hash: '0xfe15c0d3ebe314fad720a08b839a004c2e6386f5aecc19ec74807d1920cb6aeb',
+    }
+
+    await db.update(new TestBlockBuilder(block1).addTx().build());
+    await db.update(new TestBlockBuilder(block2).addTx().build());
+
     const query1 = await supertest(api.address)
       .post(`/rosetta/v1/block`)
       .send({
@@ -351,51 +305,32 @@ describe('Rosetta API', () => {
     expect(JSON.parse(query1.text)).toEqual({
       block: {
         block_identifier: {
-          index: block.result.block_height,
-          hash: block.result.block_hash,
+          index: block2.block_height,
+          hash: block2.block_hash,
         },
         parent_block_identifier: {
-          index: block.result.block_height - 1,
-          hash: block.result.parent_block_hash,
+          index: block2.block_height - 1,
+          hash: block2.parent_block_hash,
         },
-        timestamp: block.result.burn_block_time * 1000,
+        timestamp: block2.burn_block_time * 1000,
         transactions: expect.objectContaining({}),
       },
     });
   });
 
-  test.only('block/transaction', async () => {
-    const block: DbBlock = {
+  test('block/transaction', async () => {
+    const block = {
       block_hash: '0xd0dd05e3d0a1bd60640c9d9d30d57012ffe47b52fe643140c39199c757d37e3f',
       index_block_hash: '0x6a36c14514047074c2877065809bbb70d81d52507747f4616da997deb7228fad',
       parent_index_block_hash: '0x81580c80601341be11c6a2412aa342bd506a7f373fb26eda5e28d126e3429d17',
       parent_block_hash: '0x5b68076486afbb5c20730269c0ea3a0c2f26cfd200676488402a30bcdd2e8136',
       parent_microblock_hash: '0x0000000000000000000000000000000000000000000000000000000000000000',
-      parent_microblock_sequence: 0,
-      block_height: 1,
-      burn_block_time: 1646401908,
       burn_block_hash: '0xfe15c0d3ebe314fad720a08b839a004c2e6386f5aecc19ec74807d1920cb6aeb',
-      burn_block_height: 4,
       miner_txid: '0x0000000000000000000000000000000000000000000000000000000000000000',
-      canonical: true,
-      execution_cost_read_count: 0,
-      execution_cost_read_length: 0,
-      execution_cost_runtime: 0,
-      execution_cost_write_count: 0,
-      execution_cost_write_length: 0
     }
-    const tx: DbTx = {
+    const tx = {
       tx_id: '0xc152de9376bab4fc27291c9cd088643698290a12bb511d768f873cb3d280eb48',
       tx_index: 1,
-      nonce: 0,
-      raw_tx: Buffer.alloc(0),
-      index_block_hash: block.index_block_hash,
-      parent_index_block_hash: block.parent_index_block_hash,
-      block_hash: block.block_hash,
-      parent_block_hash: block.parent_block_hash,
-      block_height: 1,
-      burn_block_time: 1646401908,
-      parent_burn_block_time: 1646401907,
       type_id: DbTxTypeId.TokenTransfer,
       anchor_mode: DbTxAnchorMode.Any,
       status: DbTxStatus.Success,
@@ -406,27 +341,19 @@ describe('Rosetta API', () => {
       microblock_hash: '',
       post_conditions: Buffer.from([0x01, 0xf5]),
       fee_rate: 180n,
-      sponsored: false,
-      sponsor_address: undefined,
       sender_address: 'ST1HB1T8WRNBYB0Y3T7WXZS38NKKPTBR3EG9EPJKR',
-      origin_hash_mode: 0,
-      event_count: 1,
-      execution_cost_read_count: 0,
-      execution_cost_read_length: 0,
-      execution_cost_runtime: 0,
-      execution_cost_write_count: 0,
-      execution_cost_write_length: 0,
       abi: undefined,
       token_transfer_recipient_address: 'STRYYQQ9M8KAF4NS7WNZQYY59X93XEKR31JP64CP',
       token_transfer_amount: 3852n,
       token_transfer_memo: Buffer.from('test1234'),
     }
-    await db.update(getBlockTxUpdateData(block, tx));
+    const data = new TestBlockBuilder(block).addTx(tx).build();
+    await db.update(data);
     const query1 = await supertest(api.server)
       .post(`/rosetta/v1/block/transaction`)
       .send({
         network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { index: tx.block_height, hash: tx.block_hash },
+        block_identifier: { index: data.block.block_height, hash: data.block.block_hash },
         transaction_identifier: { hash: tx.tx_id },
       });
     expect(query1.status).toBe(200);
@@ -516,6 +443,7 @@ describe('Rosetta API', () => {
       };
       mempoolTxs.push(mempoolTx);
     }
+    
     await db.updateMempoolTxs({ mempoolTxs });
 
     const request1: RosettaMempoolRequest = {
@@ -614,6 +542,20 @@ describe('Rosetta API', () => {
   });
 
   test('account/balance success', async () => {
+    const block = {
+      block_hash: '0x6d0df2553d3dd500dec73d00e5252eec09eb6d7cd36999b7dbf011abd186bdf0',
+      index_block_hash: '0x9cd482f368644b5292bf05f5bb59a9e864386f8b96c49091ccf0818fe01c6fc5',
+    }
+
+    const stxEvent = {
+      amount: 10000000000003852n,
+      recipient: 'STRYYQQ9M8KAF4NS7WNZQYY59X93XEKR31JP64CP',
+    }
+
+    const data = new TestBlockBuilder(block).addTx().addTxStxEvent(stxEvent).build();
+
+    await db.update(data);
+
     const request1: RosettaAccountBalanceRequest = {
       network_identifier: {
         blockchain: 'stacks',
@@ -623,6 +565,7 @@ describe('Rosetta API', () => {
         address: 'STRYYQQ9M8KAF4NS7WNZQYY59X93XEKR31JP64CP',
       },
     };
+
 
     const result1 = await supertest(api.server).post(`/rosetta/v1/account/balance/`).send(request1);
     expect(result1.status).toBe(200);
@@ -641,8 +584,8 @@ describe('Rosetta API', () => {
 
     const expectedResponse: RosettaAccountBalanceResponse = {
       block_identifier: {
-        hash: curren_block.result.block_hash,
-        index: curren_block.result.block_height,
+        hash: data.block.block_hash,
+        index: data.block.block_height,
       },
       balances: [amount],
 
@@ -656,79 +599,138 @@ describe('Rosetta API', () => {
 
   test('account/balance - nonce calculated properly', async () => {
     const testAddr1 = 'STNN931GWC0XMRBWXYJQXTEKT4YFB1Z7YTCV3RZN';
-    const testAddr1Key = '532d5ff9f0d4980225a031f65a2dff75b351d675b086766917d43372cedf762901';
     const testAddr2 = 'ST2WFY0H48AS2VYPA7N69V2VJ8VKS8FSPQSPFE1Z8';
     const testAddr3 = 'ST5F760KN84TZK3VTZCTVFYCVXQBEVKNV9M7H2CW';
 
-    let expectedTxId: string = '';
-    const broadcastTx = new Promise<DbTx>(resolve => {
-      const listener: (txId: string) => void = async txId => {
-        const dbTxQuery = await api.datastore.getTx({ txId: txId, includeUnanchored: false });
-        if (!dbTxQuery.found) {
-          return;
-        }
-        const dbTx = dbTxQuery.result as DbTx;
-        if (dbTx.tx_id === expectedTxId && dbTx.status === DbTxStatus.Success) {
-          api.datastore.removeListener('txUpdate', listener);
-          resolve(dbTx);
-        }
-      };
-      api.datastore.addListener('txUpdate', listener);
-    });
-    const transferTx = await makeSTXTokenTransfer({
-      recipient: testAddr1,
-      amount: new BN(10000000),
-      senderKey: 'c71700b07d520a8c9731e4d0f095aa6efb91e16e25fb27ce2b72e7b698f8127a01',
-      network: getStacksTestnetNetwork(),
-      memo: 'test1234',
-      anchorMode: AnchorMode.Any
-    });
-    expectedTxId = '0x' + transferTx.txid();
-    const submitResult = await new StacksCoreRpcClient().sendTransaction(transferTx.serialize());
-    expect(submitResult.txId).toBe(expectedTxId);
-    let tx1 = await broadcastTx;
-    const txDb = await api.datastore.getTx({ txId: expectedTxId, includeUnanchored: false });
-    assert(txDb.found);
+    const senderAddr1 = 'ST1HB1T8WRNBYB0Y3T7WXZS38NKKPTBR3EG9EPJKR';
 
-    // Send three transactions from `testAddr1` so its nonce at chaintip should be 2 (the third nonce in a zero-based index)
-    let tx2: DbTx;
-    let tx3: DbTx;
-    let tx4: DbTx;
-    for (let i = 0; i < 3; i++) {
-      const broadcastTx2 = new Promise<DbTx>(resolve => {
-        const listener: (txId: string) => void = async txId => {
-          const dbTxQuery = await api.datastore.getTx({ txId: txId, includeUnanchored: false });
-          if (!dbTxQuery.found) {
-            return;
-          }
-          const dbTx = dbTxQuery.result as DbTx;
-          if (dbTx.tx_id === expectedTxId && dbTx.status === DbTxStatus.Success) {
-            api.datastore.removeListener('txUpdate', listener);
-            resolve(dbTx);
-          }
-        };
-        api.datastore.addListener('txUpdate', listener);
-      });
-      const transferTx2 = await makeSTXTokenTransfer({
-        recipient: testAddr2,
-        amount: new BN(10),
-        senderKey: testAddr1Key,
-        network: getStacksTestnetNetwork(),
-        memo: 'test1234',
-        anchorMode: AnchorMode.Any
-      });
-      expectedTxId = '0x' + transferTx2.txid();
-      const submitResult2 = await new StacksCoreRpcClient().sendTransaction(transferTx2.serialize());
-      expect(submitResult2.txId).toBe(expectedTxId);
-      const tx = await broadcastTx2;
-      if (i === 0) {
-        tx2 = tx;
-      } else if (i === 1) {
-        tx3 = tx;
-      } else {
-        tx4 = tx;
-      }
+    const block1 = {
+      block_hash: '0xf989ae0dd9df72ba3256737cc088579fda6c9ad3779515e1f88f6f9b6d68a4af',
+      index_block_hash: '0x2d1a360199f2992176b6c517df02e0350de863ac1d704e27f873aca41d1d5c58',
+      burn_block_time: 1647245793,
+      burn_block_hash: '0x12771355e46cd47c71ed1721fd5319b383cca3a1f9fce3aa1c8cd3bd37af20d7',
     }
+
+    const block2 = {
+      block_hash: '0x152ea79ffe71b835ccb575b732ed70354b169d44c6317752b2e1d48d15decf83',
+      index_block_hash: '0x0d5247fb314341a505fe27740d7bd84fe3f071ff6b06d3ccfb8a00ba4d51b84f',
+      parent_index_block_hash: block1.index_block_hash,
+      parent_block_hash: block1.block_hash,
+      block_height: 2,
+      burn_block_time: 1647245794,
+      burn_block_hash: '0xfe15c0d3ebe314fad720a08b839a004c2e6386f5aecc19ec74807d1920cb6aeb',
+    }
+
+    const block3 = {
+      block_hash: '0xd20083b03d0908bb7852bfbd7c9c7b5a3220dd6505b73c047da75d9a33bd59c3',
+      index_block_hash: '0xc04d68178a9181d21f11d161b0ce2d6b6d26d7e69e4ec8e6d5415f6046df6f8e',
+      parent_index_block_hash: block2.index_block_hash,
+      parent_block_hash: block2.block_hash,
+      block_height: 3,
+      burn_block_time: 1647250849,
+      burn_block_hash: '0x376da11fe3ab3d0eaaddb418ccb49b5426d5c2504f526f7766580f6e45984e3b',
+    }
+
+    const block4 = {
+      block_hash: '0x7d96609236e7cb6b55e0f3f8e94001e9b9223f5447fc97858137fed0314e0cae',
+      index_block_hash: '0x1ccfee962a39249655c9547fdf76e781e12698e0244de3f0f816f1e3d22aeb65',
+      parent_index_block_hash: block3.index_block_hash,
+      parent_block_hash: block3.block_hash,
+      block_height: 4,
+      burn_block_time: 1647250850,
+      burn_block_hash: '0x4391a5c79ffdc79883036503ca551673c09deec28df432a8d88debc7fa2ec91e',
+    }
+
+    const tx_1 =  {
+      tx_id: '0xa95f0f833cf3b90473a9654526e9cb02626b85f32efae7f05c139d3460225d5f',
+      type_id: DbTxTypeId.TokenTransfer,
+      anchor_mode: AnchorMode.Any,
+      status: DbTxStatus.Success,
+      sender_address: 'ST1HB1T8WRNBYB0Y3T7WXZS38NKKPTBR3EG9EPJKR',
+      event_count: 1,
+      token_transfer_recipient_address: testAddr1,
+      token_transfer_amount: 10000000n,
+      token_transfer_memo: Buffer.from('test1234'),
+      fee_rate: 180n,
+    }
+
+    const tx_2 = {
+      tx_id: '0xf03f217ebd89d2e0fd9bdf56ad2cc09b42455cd0d3442a4ebbb9f9dbf0dcc6b9',
+      type_id: DbTxTypeId.TokenTransfer,
+      anchor_mode: AnchorMode.Any,
+      status: DbTxStatus.Success,
+      sender_address: testAddr1,
+      event_count: 1,
+      token_transfer_recipient_address: testAddr2,
+      token_transfer_amount: 10n,
+      token_transfer_memo: Buffer.from('test1234'),
+      fee_rate: 180n,
+    }
+
+    const tx_3 = {
+      tx_id: '0xc9d72d08ae48503e93f54b218b9048b1afdb1054f15baa3244490ad41a5a5902',
+      type_id: DbTxTypeId.TokenTransfer,
+      anchor_mode: AnchorMode.Any,
+      status: DbTxStatus.Success,
+      sender_address: testAddr1,
+      event_count: 1,
+      token_transfer_recipient_address: testAddr2,
+      token_transfer_amount: 10n,
+      token_transfer_memo: Buffer.from('test1234'),
+      nonce: 1,
+      fee_rate: 180n,
+    }
+
+    const tx_4 = {
+      tx_id: '0x552e94a0d57683fd8b8e4dfa5c807096e1fa7b6ab67f651f735b9d5e3c8bea93',
+      type_id: DbTxTypeId.TokenTransfer,
+      anchor_mode: AnchorMode.Any,
+      status: DbTxStatus.Success,
+      sender_address: testAddr1,
+      event_count: 1,
+      token_transfer_recipient_address: testAddr2,
+      token_transfer_amount: 10n,
+      token_transfer_memo: Buffer.from('test1234'),
+      fee_rate: 180n,
+    }
+
+    const stx1 =  {
+      tx: tx_1,
+      amount: 10000000n,
+      sender: senderAddr1,
+      recipient: testAddr1
+    }
+
+    const stx2 = {
+      amount: 10n,
+      tx: tx_2,
+      sender: testAddr1,
+      recipient: testAddr2
+    }
+
+    const stx3 = {
+      amount: 10n,
+      tx: tx_3,
+      sender: testAddr1,
+      recipient: testAddr2,
+    }
+
+    const stx4 = {
+      amount: 10n,
+      tx: tx_4,
+      sender: testAddr1,
+      recipient: testAddr2
+    }
+
+    const data1 = new TestBlockBuilder(block1).addTx(tx_1).addTxStxEvent(stx1).build();
+    const data2 = new TestBlockBuilder(block2).addTx(tx_2).addTxStxEvent(stx2).build();
+    const data3 = new TestBlockBuilder(block3).addTx(tx_3).addTxStxEvent(stx3).build();
+    const data4 = new TestBlockBuilder(block4).addTx(tx_4).addTxStxEvent(stx4).build();
+
+    await db.update(data1);
+    await db.update(data2);
+    await db.update(data3);
+    await db.update(data4);
 
     const request1: RosettaAccountBalanceRequest = {
       network_identifier: {
@@ -736,7 +738,7 @@ describe('Rosetta API', () => {
         network: 'testnet',
       },
       block_identifier: {
-        index: tx3!.block_height,
+        index: data3.block.block_height,
       },
       account_identifier: {
         address: testAddr1,
@@ -747,8 +749,8 @@ describe('Rosetta API', () => {
     expect(nonceResult1.type).toBe('application/json');
     const expectedResponse1: RosettaAccountBalanceResponse = {
       block_identifier: {
-        hash: tx3!.block_hash,
-        index: tx3!.block_height,
+        hash: data3.block.block_hash,
+        index: data3.block.block_height,
       },
       balances: [{
         value: '9999620',
@@ -769,7 +771,7 @@ describe('Rosetta API', () => {
         network: 'testnet',
       },
       block_identifier: {
-        index: tx2!.block_height,
+        index: data2.block.block_height,
       },
       account_identifier: {
         address: testAddr1,
@@ -780,8 +782,8 @@ describe('Rosetta API', () => {
     expect(nonceResult2.type).toBe('application/json');
     const expectedResponse2: RosettaAccountBalanceResponse = {
       block_identifier: {
-        hash: tx2!.block_hash,
-        index: tx2!.block_height,
+        hash: data2.block.block_hash,
+        index: data2.block.block_height,
       },
       balances: [{
         value: '9999810',
@@ -803,7 +805,7 @@ describe('Rosetta API', () => {
         network: 'testnet',
       },
       block_identifier: {
-        index: tx2!.block_height,
+        index: data2.block.block_height,
       },
       account_identifier: {
         address: testAddr3,
@@ -814,8 +816,8 @@ describe('Rosetta API', () => {
     expect(nonceResult3.type).toBe('application/json');
     const expectedResponse3: RosettaAccountBalanceResponse = {
       block_identifier: {
-        hash: tx2!.block_hash,
-        index: tx2!.block_height,
+        hash: data2.block.block_hash,
+        index: data2.block.block_height,
       },
       balances: [{
         value: '0',
@@ -832,10 +834,39 @@ describe('Rosetta API', () => {
   });
 
   test('account/balance - fees calculated properly', async () => {
-    // this account has made one transaction
-    // ensure that the fees for it are calculated after it makes
-    // the transaction, not before, by checking its balance in block 1
     const stxAddress = 'ST1HB1T8WRNBYB0Y3T7WXZS38NKKPTBR3EG9EPJKR';
+    const senderAddr = 'STNN931GWC0XMRBWXYJQXTEKT4YFB1Z7YTCV3RZN';
+
+    const block1 = {
+      block_hash: '0xf989ae0dd9df72ba3256737cc088579fda6c9ad3779515e1f88f6f9b6d68a4af',
+      index_block_hash: '0x2d1a360199f2992176b6c517df02e0350de863ac1d704e27f873aca41d1d5c58',
+      burn_block_time: 1647245793,
+      burn_block_hash: '0x12771355e46cd47c71ed1721fd5319b383cca3a1f9fce3aa1c8cd3bd37af20d7',
+    }
+
+    const tx_1 =  {
+      tx_id: '0xa95f0f833cf3b90473a9654526e9cb02626b85f32efae7f05c139d3460225d5f',
+      type_id: DbTxTypeId.TokenTransfer,
+      anchor_mode: AnchorMode.Any,
+      status: DbTxStatus.Success,
+      sender_address: senderAddr,
+      event_count: 1,
+      token_transfer_recipient_address: stxAddress,
+      token_transfer_amount: 10000000000000000n,
+      token_transfer_memo: Buffer.from('test1234'),
+    }
+
+    const stx1 =  {
+      tx: tx_1,
+      amount: tx_1.token_transfer_amount,
+      sender: senderAddr,
+      recipient: stxAddress
+    }
+
+    const data1 = new TestBlockBuilder(block1).addTx(tx_1).addTxStxEvent(stx1).build();
+
+    await db.update(data1);
+
     const request1: RosettaAccountBalanceRequest = {
       network_identifier: {
         blockchain: 'stacks',
@@ -853,9 +884,6 @@ describe('Rosetta API', () => {
     expect(result1.status).toBe(200);
     expect(result1.type).toBe('application/json');
 
-    const block = await api.datastore.getBlock({ height: 1 });
-    assert(block.found);
-
     const amount: RosettaAmount = {
       value: '10000000000000000',
       currency: {
@@ -866,8 +894,8 @@ describe('Rosetta API', () => {
 
     const expectedResponse: RosettaAccountBalanceResponse = {
       block_identifier: {
-        hash: block.result.block_hash,
-        index: block.result.block_height,
+        hash: data1.block.block_hash,
+        index: data1.block.block_height,
       },
       balances: [amount],
 
@@ -957,2564 +985,7 @@ describe('Rosetta API', () => {
     expect(JSON.parse(result.text)).toEqual(expectResponse);
   });
 
-  /* rosetta construction api tests below */
-
-  test('construction/derive', async () => {
-    const request: RosettaConstructionDeriveRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      public_key: {
-        curve_type: 'secp256k1',
-        hex_bytes: publicKeyToString(
-          getPublicKey(createStacksPrivateKey(testnetKeys[0].secretKey))
-        ),
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/derive`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const accountIdentifier: RosettaAccountIdentifier = {
-      address: testnetKeys[0].stacksAddress,
-    };
-    const expectResponse: RosettaConstructionDeriveResponse = {
-      account_identifier: accountIdentifier,
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-
-    const request2 = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      public_key: {
-        curve_type: 'this is an invalid curve type',
-        hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-      },
-    };
-
-    const result2 = await supertest(api.server)
-      .post(`/rosetta/v1/construction/derive`)
-      .send(request2);
-    expect(result2.status).toBe(400);
-
-    const expectedResponse2 = RosettaErrors[RosettaErrorsTypes.invalidCurveType];
-
-    expect(JSON.parse(result2.text)).toEqual(expectedResponse2);
-
-    const request3 = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      public_key: {
-        curve_type: 'secp256k1',
-        hex_bytes: 'this is an invalid public key',
-      },
-    };
-
-    const result3 = await supertest(api.server)
-      .post(`/rosetta/v1/construction/derive`)
-      .send(request3);
-    expect(result3.status).toBe(400);
-
-    const expectedResponse3 = RosettaErrors[RosettaErrorsTypes.invalidPublicKey];
-
-    expect(JSON.parse(result3.text)).toEqual(expectedResponse3);
-  });
-
-  test('construction/preprocess - Stack STX', async () => {
-    const sender = testnetKeys[0].stacksAddress;
-    const stacking_amount = 5000;
-    const number_of_cycles = 5;
-
-    const request: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'stack_stx',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-' + stacking_amount,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-          },
-          metadata: {
-            number_of_cycles: number_of_cycles,
-            pox_addr : '1Xik14zRm29UsyS6DjhYg4iZeZqsDa8D3',
-          }
-        },
-      ],
-      metadata: {
-        memo: 'SAMPLE MEMO',
-      },
-      max_fee: [
-        {
-          value: '12380898',
-          currency: {
-            symbol: 'STX',
-            decimals: 6,
-          },
-          metadata: {},
-        },
-      ],
-      suggested_fee_multiplier: 1,
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/preprocess`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const expectResponse: RosettaConstructionPreprocessResponse = {
-      options: {
-        sender_address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        type: 'stack_stx',
-        number_of_cycles: 5,
-        amount: '5000',
-        symbol: 'STX',
-        decimals: 6,
-        pox_addr: '1Xik14zRm29UsyS6DjhYg4iZeZqsDa8D3',
-        suggested_fee_multiplier: 1,
-        max_fee: '12380898',
-        size: 260,
-        memo: 'SAMPLE MEMO'
-      },
-      required_public_keys: [ { address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6' } ]
-    }
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-  });
-
-  test('construction/preprocess - token transfer', async () => {
-    const request: RosettaConstructionPreprocessRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-            metadata: {},
-          },
-          amount: {
-            value: '-500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-            metadata: {},
-          },
-          amount: {
-            value: '500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-      ],
-      metadata: {
-        memo: 'SAMPLE MEMO',
-      },
-      max_fee: [
-        {
-          value: '12380898',
-          currency: {
-            symbol: 'STX',
-            decimals: 6,
-          },
-          metadata: {},
-        },
-      ],
-      suggested_fee_multiplier: 1,
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/preprocess`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const expectResponse: RosettaConstructionPreprocessResponse = {
-      options: {
-        sender_address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        type: 'token_transfer',
-        suggested_fee_multiplier: 1,
-        token_transfer_recipient_address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        memo: 'SAMPLE MEMO',
-        size: 180,
-      },
-      required_public_keys: [
-        {
-          address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        },
-      ],
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-  });
-
-  test('construction/preprocess - failure', async () => {
-    const request2 = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'invalid operation type',
-          account: {
-            address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-            metadata: {},
-          },
-          amount: {
-            value: '-500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-            metadata: {},
-          },
-          amount: {
-            value: '500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-      ],
-      metadata: {},
-      max_fee: [
-        {
-          value: '12380898',
-          currency: {
-            symbol: 'STX',
-            decimals: 6,
-          },
-          metadata: {},
-        },
-      ],
-      suggested_fee_multiplier: 1,
-    };
-
-    const result2 = await supertest(api.server)
-      .post(`/rosetta/v1/construction/preprocess`)
-      .send(request2);
-    expect(result2.status).toBe(400);
-
-    const expectedResponse2 = RosettaErrors[RosettaErrorsTypes.invalidOperation];
-
-    expect(JSON.parse(result2.text)).toEqual(expectedResponse2);
-  });
-
-  test('construction/metadata - success', async () => {
-    const request: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      options: {
-        sender_address: testnetKeys[0].stacksAddress,
-        type: 'token_transfer',
-        suggested_fee_multiplier: 1,
-        token_transfer_recipient_address: testnetKeys[1].stacksAddress,
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        size: 180,
-        memo: 'SAMPLE MEMO',
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-    expect(JSON.parse(result.text)).toHaveProperty('metadata');
-    expect(JSON.parse(result.text)).toHaveProperty('suggested_fee');
-    expect(JSON.parse(result.text).suggested_fee[0].value).toBe('180');
-    expect(JSON.parse(result.text).metadata.memo).toBe('SAMPLE MEMO');
-  });
-
-  test('construction/metadata - empty network identifier', async () => {
-    const request = {
-      options: {
-        sender_address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        type: 'token_transfer',
-        token_transfer_recipient_address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        size: 180,
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectResponse = {
-      code: 613,
-      message: 'Network identifier object is null.',
-      retriable: false,
-      details: {
-        message: "should have required property 'network_identifier'",
-      },
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-  });
-
-  test('construction/metadata - invalid transfer type', async () => {
-    const request: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      options: {
-        sender_address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        type: 'token',
-        token_transfer_recipient_address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        size: 180,
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectResponse = {
-      code: 625,
-      message: 'Invalid transaction type',
-      retriable: false,
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-  });
-
-  test('construction/metadata - invalid sender address', async () => {
-    const request: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      options: {
-        sender_address: 'abc',
-        type: 'token_transfer',
-        token_transfer_recipient_address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        fee: '-180',
-        max_fee: '12380898',
-        size: 180,
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectResponse = {
-      code: 626,
-      message: 'Invalid sender address',
-      retriable: false,
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-  });
-
-  test('construction/metadata - invalid recipient address', async () => {
-    const request: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      options: {
-        sender_address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        type: 'token_transfer',
-        token_transfer_recipient_address: 'xyz',
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        size: 180,
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectResponse = {
-      code: 627,
-      message: 'Invalid recipient address',
-      retriable: false,
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectResponse);
-  });
-
-  test('construction/hash', async () => {
-    const request: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      signed_transaction:
-        '0x80800000000400539886f96611ba3ba6cef9618f8c78118b37c5be000000000000000000000000000000b400017a33a91515ef48608a99c6adecd2eb258e11534a1acf66348f5678c8e2c8f83d243555ed67a0019d3500df98563ca31321c1a675b43ef79f146e322fe08df75103020000000000051a1ae3f911d8f1d46d7416bfbe4b593fd41eac19cb000000000007a12000000000000000000000000000000000000000000000000000000000000000000000',
-    };
-
-    const result = await supertest(api.server).post(`/rosetta/v1/construction/hash`).send(request);
-    expect(result.status).toBe(200);
-
-    const expectedResponse: RosettaConstructionHashResponse = {
-      transaction_identifier: {
-        hash: '0xf3b054a5fbae98f7f35e5e917b65759fc365a3e073f8af1c3b8d211b286fa74a',
-      },
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('construction/hash - no `0x` prefix', async () => {
-    const request: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      signed_transaction:
-        '0x80800000000400d429e0b599f9cba40ecc9f219df60f9d0a02212d000000000000000100000000000000000101cc0235071690bc762d0013f6d3e4be32aa8f8d01d0db9d845595589edba47e7425bd655f20398e3d931cbe60eea59bb66f44d3f28443078fe9d10082dccef80c010200000000040000000000000000000000000000000000000000000000000000000000000000',
-    };
-
-    const result = await supertest(api.server).post(`/rosetta/v1/construction/hash`).send(request);
-    expect(result.status).toBe(200);
-
-    const expectedResponse: RosettaConstructionHashResponse = {
-      transaction_identifier: {
-        hash: '0x592fad4733f3e5c65e7dd9c82ad848191993a80cb3d891b6514bda6e3e7a239e',
-      },
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('construction/hash - odd number of hex digits', async () => {
-    const request: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      signed_transaction:
-        '80800000000400d429e0b599f9cba40ecc9f219df60f9d0a02212d000000000000000100000000000000000101cc0235071690bc762d0013f6d3e4be32aa8f8d01d0db9d845595589edba47e7425bd655f20398e3d931cbe60eea59bb66f44d3f28443078fe9d10082dccef80c01020000000004000000000000000000000000000000000000000000000000000000000000000',
-    };
-
-    const result = await supertest(api.server).post(`/rosetta/v1/construction/hash`).send(request);
-    expect(result.status).toBe(400);
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.invalidTransactionString];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('construction/hash - unsigned transaction', async () => {
-    const request: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      //unsigned transaction bytes
-      signed_transaction:
-        '0x80800000000400539886f96611ba3ba6cef9618f8c78118b37c5be000000000000000000000000000000b400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003020000000000051a1ae3f911d8f1d46d7416bfbe4b593fd41eac19cb000000000007a12000000000000000000000000000000000000000000000000000000000000000000000',
-    };
-
-    const result = await supertest(api.server).post(`/rosetta/v1/construction/hash`).send(request);
-    expect(result.status).toBe(400);
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.transactionNotSigned];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('construction/parse - signed', async () => {
-    const publicKey = publicKeyToString(
-      getPublicKey(createStacksPrivateKey(testnetKeys[0].secretKey))
-    );
-    const senderAddr = testnetKeys[0].stacksAddress;
-    const recipientAddr = 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0';
-    const amount = new BN(1000);
-    const fee = new BN(180);
-    const options: SignedTokenTransferOptions = {
-      recipient: recipientAddr,
-      amount: amount,
-      fee: fee,
-      senderKey: testnetKeys[0].secretKey,
-      network: getStacksTestnetNetwork(),
-      anchorMode: AnchorMode.Any,
-    };
-    const testTransaction = await makeSTXTokenTransfer(options);
-    const request: RosettaConstructionParseRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      signed: true,
-      transaction: bufferToHexPrefixString(testTransaction.serialize()),
-    };
-
-    const result = await supertest(api.server).post(`/rosetta/v1/construction/parse`).send(request);
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const actual: RosettaConstructionParseResponse = JSON.parse(result.text);
-    // test fee operation
-    expect(actual.operations[0].account?.address).toEqual(senderAddr);
-    expect(actual.operations[0].amount?.value).toEqual('-' + fee.toString());
-    // test sender
-    expect(actual.operations[1].account?.address).toEqual(senderAddr);
-    expect(actual.operations[1].amount?.value).toEqual('-' + amount.toString());
-    // test recipient
-    expect(actual.operations[2].account?.address).toEqual(recipientAddr);
-    expect(actual.operations[2].amount?.value).toEqual(amount.toString());
-    // test signer
-    expect(actual.account_identifier_signers?.[0].address).toEqual(senderAddr);
-  });
-
-  test('construction/parse - unsigned', async () => {
-    const publicKey = publicKeyToString(
-      getPublicKey(createStacksPrivateKey(testnetKeys[0].secretKey))
-    );
-    const senderAddr = testnetKeys[0].stacksAddress;
-    const recipientAddr = 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0';
-    const amount = new BN(1000);
-    const fee = new BN(180);
-    const tokenTransferOptions: UnsignedTokenTransferOptions = {
-      recipient: recipientAddr,
-      amount: amount,
-      fee: fee,
-      publicKey: publicKey,
-      network: getStacksTestnetNetwork(),
-      anchorMode: AnchorMode.Any,
-    };
-    const testTransaction = await makeUnsignedSTXTokenTransfer(tokenTransferOptions);
-
-    const request: RosettaConstructionParseRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      signed: false,
-      transaction: bufferToHexPrefixString(testTransaction.serialize()),
-    };
-
-    const result = await supertest(api.server).post(`/rosetta/v1/construction/parse`).send(request);
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-    const actual: RosettaConstructionParseResponse = JSON.parse(result.text);
-    // test fee operation
-    expect(actual.operations[0].account?.address).toEqual(senderAddr);
-    expect(actual.operations[0].amount?.value).toEqual('-' + fee.toString());
-    // test sender
-    expect(actual.operations[1].account?.address).toEqual(senderAddr);
-    expect(actual.operations[1].amount?.value).toEqual('-' + amount.toString());
-    // test recipient
-    expect(actual.operations[2].account?.address).toEqual(recipientAddr);
-    expect(actual.operations[2].amount?.value).toEqual(amount.toString());
-  });
-
-  test('construction/submit', async () => {
-    const txOptions = {
-      senderKey: testnetKeys[0].secretKey,
-      recipient: standardPrincipalCV(testnetKeys[1].stacksAddress),
-      amount: new BigNum(12345),
-      network: getStacksTestnetNetwork(),
-      memo: 'test memo',
-      nonce: new BigNum(0),
-      fee: new BigNum(200),
-      anchorMode: AnchorMode.Any,
-    };
-
-    const transaction = await makeSTXTokenTransfer(txOptions);
-    const serializedTx = transaction.serialize().toString('hex');
-
-    const request: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      // signed transaction bytes
-      signed_transaction: '0x' + serializedTx,
-    };
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/submit`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-  });
-
-  test('construction/submit - unsigned', async () => {
-    const txOptions = {
-      recipient: standardPrincipalCV(testnetKeys[1].stacksAddress),
-      amount: new BigNum(12345),
-      publicKey: publicKeyToString(pubKeyfromPrivKey(testnetKeys[0].secretKey)),
-      network: getStacksTestnetNetwork(),
-      memo: 'test memo',
-      nonce: new BigNum(0),
-      fee: new BigNum(200),
-      anchorMode: AnchorMode.Any,
-    };
-
-    const transaction = await makeUnsignedSTXTokenTransfer(txOptions);
-    const serializedTx = transaction.serialize().toString('hex');
-
-    const request: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      //unsigned transaction bytes
-      signed_transaction: '0x' + serializedTx,
-    };
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/submit`)
-      .send(request);
-    expect(result.status).toBe(400);
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.invalidTransactionString];
-
-    expect(JSON.parse(result.text)).toMatchObject(expectedResponse);
-  });
-
-  test('payloads single sign success', async () => {
-    const publicKey = publicKeyToString(pubKeyfromPrivKey(testnetKeys[0].secretKey));
-    const sender = testnetKeys[0].stacksAddress;
-    const recipient = testnetKeys[1].stacksAddress;
-    const fee = '180';
-
-    const request: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-' + fee,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 2,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: recipient,
-            metadata: {},
-          },
-          amount: {
-            value: '500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-      ],
-      metadata: {
-        account_sequence: 0,
-        memo: 'SAMPLE MEMO',
-      },
-      public_keys: [
-        {
-          hex_bytes: publicKey,
-          curve_type: 'secp256k1',
-        },
-      ],
-    };
-
-    const tokenTransferOptions: UnsignedTokenTransferOptions = {
-      recipient: recipient,
-      amount: new BN('500000'),
-      fee: new BN(fee),
-      publicKey: publicKey,
-      network: getStacksNetwork(),
-      nonce: new BN(0),
-      memo: 'SAMPLE MEMO',
-      anchorMode: AnchorMode.Any,
-    };
-
-    const transaction = await makeUnsignedSTXTokenTransfer(tokenTransferOptions);
-    const unsignedTransaction = transaction.serialize();
-    // const hexBytes = digestSha512_256(unsignedTransaction).toString('hex');
-
-    const signer = new TransactionSigner(transaction);
-
-    const prehash = makeSigHashPreSign(signer.sigHash, AuthType.Standard, new BN(fee), new BN(0));
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const accountIdentifier: RosettaAccountIdentifier = {
-      address: sender,
-    };
-
-    const expectedResponse = {
-      unsigned_transaction: '0x' + unsignedTransaction.toString('hex'),
-      payloads: [
-        {
-          address: sender,
-          account_identifier: accountIdentifier,
-          hex_bytes: prehash,
-          signature_type: 'ecdsa_recovery',
-        },
-      ],
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('payloads multi sig', async () => {
-    const publicKey1 = publicKeyToString(pubKeyfromPrivKey(testnetKeys[0].secretKey));
-    const publicKey2 = publicKeyToString(pubKeyfromPrivKey(testnetKeys[1].secretKey));
-
-    const sender = testnetKeys[0].stacksAddress;
-    const recipient = testnetKeys[1].stacksAddress;
-    const fee = '180';
-
-    const request: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-' + fee,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 2,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: recipient,
-            metadata: {},
-          },
-          amount: {
-            value: '500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-      ],
-      metadata: {
-        fee,
-        account_sequence: 0,
-      },
-      public_keys: [
-        {
-          hex_bytes: publicKey1,
-          curve_type: 'secp256k1',
-        },
-        {
-          hex_bytes: publicKey2,
-          curve_type: 'secp256k1',
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.needOnePublicKey];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('payloads single sign - stacking', async () => {
-    const publicKey = publicKeyToString(pubKeyfromPrivKey(testnetKeys[0].secretKey));
-    const sender = testnetKeys[0].stacksAddress;
-    const fee = '180';
-    const contract_address = 'ST000000000000000000002AMW42H';
-    const contract_name = 'pox';
-    const stacking_amount = 5000;
-    const burn_block_height = 200;
-    const number_of_cycles = 5;
-
-    const request: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: fee,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata:{
-            	number_of_cycles: 5
-            },
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'stack_stx',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-' + stacking_amount,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-          },
-          metadata: {
-            number_of_cycles: number_of_cycles,
-            pox_addr : '1Xik14zRm29UsyS6DjhYg4iZeZqsDa8D3',
-
-          }
-        },
-      ],
-      metadata: {
-        account_sequence: 0,
-        number_of_cycles: number_of_cycles, 
-        contract_address: contract_address, 
-        contract_name: contract_name,
-        burn_block_height: burn_block_height, 
-
-      },
-      public_keys: [
-        {
-          hex_bytes: publicKey,
-          curve_type: 'secp256k1',
-        },
-      ],
-    };
-
-    const poxBTCAddress = '1Xik14zRm29UsyS6DjhYg4iZeZqsDa8D3'
-
-    const { hashMode, data } = decodeBtcAddress(poxBTCAddress);
-    const hashModeBuffer = bufferCV(new BN(hashMode, 10).toArrayLike(Buffer));
-    const hashbytes = bufferCV(data);
-    const poxAddressCV = tupleCV({
-      hashbytes,
-      version: hashModeBuffer,
-    });
-
-
-    const stackingTx: UnsignedContractCallOptions = {
-      contractAddress: contract_address,
-      contractName: contract_name,
-      functionName: 'stack-stx',
-      publicKey: publicKey,
-      functionArgs: [
-        uintCV(stacking_amount),
-        poxAddressCV,
-        uintCV(burn_block_height),
-        uintCV(number_of_cycles),
-      ],
-      validateWithAbi: false,
-      nonce: new BN(0), 
-      fee: new BN(fee),
-      network: getStacksNetwork(),
-      anchorMode: AnchorMode.Any,
-    };
-    const transaction = await makeUnsignedContractCall(stackingTx);
-    const unsignedTransaction = transaction.serialize();
-    // const hexBytes = digestSha512_256(unsignedTransaction).toString('hex');
-
-    const signer = new TransactionSigner(transaction);
-
-    const prehash = makeSigHashPreSign(signer.sigHash, AuthType.Standard, new BN(fee), new BN(0));
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const accountIdentifier: RosettaAccountIdentifier = {
-      address: sender,
-    };
-
-    const expectedResponse = {
-      unsigned_transaction: '0x' + unsignedTransaction.toString('hex'),
-      payloads: [
-        {
-          address: sender,
-          account_identifier: accountIdentifier,
-          hex_bytes: prehash,
-          signature_type: 'ecdsa_recovery',
-        },
-      ],
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('payloads public key not added', async () => {
-    const request: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-            metadata: {},
-          },
-          amount: {
-            value: '-180',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-            metadata: {},
-          },
-          amount: {
-            value: '-500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 2,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-            metadata: {},
-          },
-          amount: {
-            value: '500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-      ],
-      metadata: {
-        fee: '180',
-        account_sequence: 0,
-      },
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.emptyPublicKey];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('payloads public key invalid curve type', async () => {
-    const request: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-            metadata: {},
-          },
-          amount: {
-            value: '-180',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-            metadata: {},
-          },
-          amount: {
-            value: '-500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-        {
-          operation_identifier: {
-            index: 2,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'token_transfer',
-          account: {
-            address: 'STDE7Y8HV3RX8VBM2TZVWJTS7ZA1XB0SSC3NEVH0',
-            metadata: {},
-          },
-          amount: {
-            value: '500000',
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-        },
-      ],
-      metadata: {
-        account_sequence: 0,
-      },
-      public_keys: [
-        {
-          hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-          curve_type: 'edwards25519',
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.invalidCurveType];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('combine single sign success', async () => {
-    const publicKey = publicKeyToString(pubKeyfromPrivKey(testnetKeys[0].secretKey));
-
-    const txOptions: UnsignedTokenTransferOptions = {
-      publicKey: publicKey,
-      recipient: standardPrincipalCV(testnetKeys[1].stacksAddress),
-      amount: new BigNum(12345),
-      network: getStacksTestnetNetwork(),
-      memo: 'test memo',
-      nonce: new BigNum(0),
-      fee: new BigNum(200),
-      anchorMode: AnchorMode.Any,
-    };
-
-    const unsignedTransaction = await makeUnsignedSTXTokenTransfer(txOptions);
-    const unsignedSerializedTx = unsignedTransaction.serialize().toString('hex');
-
-    const signer = new TransactionSigner(unsignedTransaction);
-
-    const prehash = makeSigHashPreSign(signer.sigHash, AuthType.Standard, new BigNum(200), new BigNum(0));
-
-    signer.signOrigin(createStacksPrivateKey(testnetKeys[0].secretKey));
-    const signedSerializedTx = signer.transaction.serialize().toString('hex');
-
-    const signature: MessageSignature = getSignature(signer.transaction) as MessageSignature;
-
-    const request: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      unsigned_transaction: '0x' + unsignedSerializedTx,
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes: prehash,
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: publicKey,
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes: signature.data.slice(2) + signature.data.slice(0, 2),
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse: RosettaConstructionCombineResponse = {
-      signed_transaction: '0x' + signedSerializedTx,
-    };
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('combine multi sig', async () => {
-    const request: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      unsigned_transaction:
-        '00000000010400539886f96611ba3ba6cef9618f8c78118b37c5be0000000000000000000000000000006400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003020000000000051ab71a091b4b8b7661a661c620966ab6573bc2dcd3000000000007a12074657374207472616e73616374696f6e000000000000000000000000000000000000',
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes:
-              '017c7fc676effda9d905440a052d304b5d9705c30625e654f5b3c9ed461337cdec695d14e5f24091d61f8409f2ab703102ca840dbf5ef752ec534fe1f418552201',
-            signature_type: 'ecdsa',
-          },
-          public_key: {
-            hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa',
-          hex_bytes:
-            '017c7fc676effda9d905440a052d304b5d9705c30625e654f5b3c9ed461337cdec695d14e5f24091d61f8409f2ab703102ca840dbf5ef752ec534fe1f418552201',
-        },
-        {
-          signing_payload: {
-            hex_bytes:
-              '017c7fc676effda9d905440a052d304b5d9705c30625e654f5b3c9ed461337cdec695d14e5f24091d61f8409f2ab703102ca840dbf5ef752ec534fe1f418552201',
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes:
-            '017c7fc676effda9d905440a052d304b5d9705c30625e654f5b3c9ed461337cdec695d14e5f24091d61f8409f2ab703102ca840dbf5ef752ec534fe1f418552201',
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.needOnlyOneSignature];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('combine invalid transaction', async () => {
-    const request: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      unsigned_transaction: 'invalid transaction',
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes:
-              '36212600bf7463399a23c398f29ca7006b9986b4a01129dd7c6e89314607208e516b0b28c1d850fe6e164abea7b6cceb4aa09700a6d218d1b605d4a402d3038f01',
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes:
-            '36212600bf7463399a23c398f29ca7006b9986b4a01129dd7c6e89314607208e516b0b28c1d850fe6e164abea7b6cceb4aa09700a6d218d1b605d4a402d3038f01',
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.invalidTransactionString];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('combine invalid signature', async () => {
-    const request: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      unsigned_transaction:
-        '00000000010400539886f96611ba3ba6cef9618f8c78118b37c5be0000000000000000000000000000006400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003020000000000051ab71a091b4b8b7661a661c620966ab6573bc2dcd3000000000007a12074657374207472616e73616374696f6e000000000000000000000000000000000000',
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes: 'invalid signature',
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes: 'invalid signature',
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.invalidSignature];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('combine signature not verified', async () => {
-    const publicKey = publicKeyToString(pubKeyfromPrivKey(testnetKeys[0].secretKey));
-
-    const txOptions: UnsignedTokenTransferOptions = {
-      publicKey: publicKey,
-      recipient: standardPrincipalCV(testnetKeys[1].stacksAddress),
-      amount: new BigNum(12345),
-      network: getStacksTestnetNetwork(),
-      memo: 'test memo',
-      nonce: new BigNum(0),
-      fee: new BigNum(200),
-      anchorMode: AnchorMode.Any,
-    };
-
-    const unsignedTransaction = await makeUnsignedSTXTokenTransfer(txOptions);
-    const unsignedSerializedTx = unsignedTransaction.serialize().toString('hex');
-
-    const signer = new TransactionSigner(unsignedTransaction);
-    signer.signOrigin(createStacksPrivateKey(testnetKeys[1].secretKey)); // use different secret key to sign
-
-    const signature = getSignature(unsignedTransaction);
-    if (!signature) throw Error('Signature undefined');
-
-    const request: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      unsigned_transaction: unsignedSerializedTx,
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes: signature.data,
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: '025c13b2fc2261956d8a4ad07d481b1a3b2cbf93a24f992249a61c3a1c4de79c51',
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes: signature.data,
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.signatureNotVerified];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-
-    // testing undefined signature
-    const incorrectTransaction = unsignedTransaction;
-    delete incorrectTransaction.auth.spendingCondition;
-    const undefinedSignature = getSignature(incorrectTransaction);
-    expect(undefinedSignature).toBe(undefined);
-
-    // testing undefined pre sign hash
-    delete incorrectTransaction.auth.authType;
-    const undefinedPreSignHash = makePresignHash(incorrectTransaction);
-    expect(undefinedPreSignHash).toBe(undefined);
-  });
-
-  test('combine invalid public key', async () => {
-    const request: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      unsigned_transaction:
-        '80800000000400539886f96611ba3ba6cef9618f8c78118b37c5be000000000000000000000000000000b4000136212600bf7463399a23c398f29ca7006b9986b4a01129dd7c6e89314607208e516b0b28c1d850fe6e164abea7b6cceb4aa09700a6d218d1b605d4a402d3038f03020000000000051ab71a091b4b8b7661a661c620966ab6573bc2dcd3000000000007a12074657374207472616e73616374696f6e000000000000000000000000000000000000',
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes:
-              '36212600bf7463399a23c398f29ca7006b9986b4a01129dd7c6e89314607208e516b0b28c1d850fe6e164abea7b6cceb4aa09700a6d218d1b605d4a402d3038f01',
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: 'invalid  public key',
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes:
-            '36212600bf7463399a23c398f29ca7006b9986b4a01129dd7c6e89314607208e516b0b28c1d850fe6e164abea7b6cceb4aa09700a6d218d1b605d4a402d3038f01',
-        },
-      ],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(request);
-
-    expect(result.status).toBe(400);
-    expect(result.type).toBe('application/json');
-
-    const expectedResponse = RosettaErrors[RosettaErrorsTypes.signatureNotVerified];
-
-    expect(JSON.parse(result.text)).toEqual(expectedResponse);
-  });
-
-  test('construction/metadata - stacking', async () => {
-    const publicKey = publicKeyToString(
-      getPublicKey(createStacksPrivateKey(testnetKeys[0].secretKey))
-    );
-    const request: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      options: {
-        sender_address: 'STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6',
-        type: 'stack_stx',
-        suggested_fee_multiplier: 1,
-        amount: '-500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        number_of_cycles: 3,
-        size: 260,
-      },
-      public_keys: [{ hex_bytes: publicKey, curve_type: 'secp256k1' }],
-    };
-
-    await standByForPoxToBeReady();
-
-    const result = await supertest(api.server)
-    .post(`/rosetta/v1/construction/metadata`)
-    .send(request);
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    expect(JSON.parse(result.text)).toHaveProperty('metadata');
-    expect(JSON.parse(result.text)).toHaveProperty('suggested_fee');
-    expect(JSON.parse(result.text).metadata).toHaveProperty('contract_address');
-    expect(JSON.parse(result.text).metadata).toHaveProperty('contract_name');
-    expect(JSON.parse(result.text).metadata).toHaveProperty('burn_block_height');
-    expect(JSON.parse(result.text).suggested_fee[0].value).toBe('260');
-
-  });
-
-  test('construction/metadata - delegate_stacking', async () => {
-    const publicKey = publicKeyToString(
-      getPublicKey(createStacksPrivateKey(testnetKeys[0].secretKey))
-    );
-    const request: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: 'stacks',
-        network: 'testnet',
-      },
-      options: {
-        fee: '180',
-        sender_address: testnetKeys[0].stacksAddress,
-        type: 'delegate_stx',
-        suggested_fee_multiplier: 1,
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        delegate_to: testnetKeys[1].stacksAddress,
-        size: 260,
-      },
-      public_keys: [{ hex_bytes: publicKey, curve_type: 'secp256k1' }],
-    };
-
-    const result = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(request);
-
-    expect(result.status).toBe(200);
-    expect(result.type).toBe('application/json');
-
-    const accountInfo = await new StacksCoreRpcClient().getAccount(testnetKeys[0].stacksAddress);
-    const nonce = accountInfo.nonce;
-
-    const metadataResponse: RosettaConstructionMetadataResponse = {
-      metadata: {
-        fee: '180',
-        sender_address: testnetKeys[0].stacksAddress,
-        type: 'delegate_stx',
-        suggested_fee_multiplier: 1,
-        amount: '500000',
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: '12380898',
-        delegate_to: testnetKeys[1].stacksAddress,
-        size: 260,
-        contract_address: 'ST000000000000000000002AMW42H',
-        contract_name: 'pox',
-        account_sequence: nonce,
-        recent_block_hash: '0x969e494d5aee0166016836f97bbeb3d9473bea8427e477e9de253f78d3212354'
-      },
-      suggested_fee: [ { value: '260', currency: {symbol: 'STX', decimals: 6} } ]
-    }
-
-    expect(result.body).toHaveProperty('metadata');
-    expect(result.body.suggested_fee).toStrictEqual(metadataResponse.suggested_fee);
-    expect(result.body.metadata.sender_address).toBe(metadataResponse.metadata.sender_address);
-    expect(result.body.metadata.type).toBe(metadataResponse.metadata.type);
-    expect(result.body.metadata.suggested_fee_multiplier).toBe(metadataResponse.metadata.suggested_fee_multiplier);
-    expect(result.body.metadata.amount).toBe(metadataResponse.metadata.amount);
-    expect(result.body.metadata.symbol).toBe(metadataResponse.metadata.symbol);
-    expect(result.body.metadata.decimals).toBe(metadataResponse.metadata.decimals);
-    expect(result.body.metadata.max_fee).toBe(metadataResponse.metadata.max_fee);
-    expect(result.body.metadata.delegate_to).toBe(metadataResponse.metadata.delegate_to);
-    expect(result.body.metadata.size).toBe(metadataResponse.metadata.size);
-    expect(result.body.metadata.contract_address).toBe(metadataResponse.metadata.contract_address);
-    expect(result.body.metadata.contract_name).toBe(metadataResponse.metadata.contract_name);
-    expect(result.body.metadata.fee).toBe(metadataResponse.metadata.fee);
-    expect(result.body.metadata.nonce).toBe(metadataResponse.metadata.nonce);
-    expect(result.body.metadata.recent_block_hash).toBeTruthy();
-  });
-
-  test('stacking rosetta transaction cycle', async() => {
-
-    //derive
-    const publicKey = publicKeyToString(
-      getPublicKey(createStacksPrivateKey(testnetKeys[0].secretKey))
-    );
-    const deriveRequest: RosettaConstructionDeriveRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      public_key: {
-        curve_type: 'secp256k1',
-        hex_bytes: publicKey,
-      },
-    };
-    const deriveResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/derive`)
-      .send(deriveRequest);
-
-    expect(deriveResult.status).toBe(200);
-    expect(deriveResult.type).toBe('application/json');
-
-    const deriveExpectResponse: RosettaConstructionDeriveResponse = {
-      account_identifier: { address: testnetKeys[0].stacksAddress },
-    };
-
-    expect(deriveResult.body).toEqual(deriveExpectResponse);
-
-
-    //preprocess
-    const fee = '260';
-    const stacking_amount = '1250180000000000';//minimum stacking 
-    const sender = deriveResult.body.account_identifier.address;
-    const number_of_cycles = 3;
-    const pox_addr = '2MtzNEqm2D9jcbPJ5mW7Z3AUNwqt3afZH66';
-    const size = 260;
-    const max_fee = '12380898';
-    const preprocessRequest: RosettaConstructionPreprocessRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: [
-
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'stack_stx',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-'+stacking_amount,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-          metadata: {
-            number_of_cycles: number_of_cycles,
-            pox_addr: pox_addr,
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: fee,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-          },
-        }
-      ],
-      metadata: {},
-      max_fee: [
-        {
-          value: max_fee,
-          currency: {
-            symbol: 'STX',
-            decimals: 6,
-          },
-          metadata: {},
-        },
-      ],
-      suggested_fee_multiplier: 1,
-    };
-    const preprocessResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/preprocess`)
-      .send(preprocessRequest);
-
-    expect(preprocessResult.status).toBe(200);
-    expect(preprocessResult.type).toBe('application/json');
-
-    const expectResponse: RosettaConstructionPreprocessResponse = {
-      options: {
-        fee: fee,
-        sender_address: sender,
-        type: 'stack_stx',
-        suggested_fee_multiplier: 1,
-        amount: stacking_amount,
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: max_fee,
-        size: size,
-        number_of_cycles: number_of_cycles,
-        pox_addr: pox_addr
-      },
-      required_public_keys: [
-        {
-          address: sender,
-        },
-      ],
-    };
-    expect(preprocessResult.body).toEqual(expectResponse);
-
-    //metadata 
-    const metadataRequest: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      options: preprocessResult.body.options, //using options returned from preprocess 
-      public_keys: [{ hex_bytes: publicKey, curve_type: 'secp256k1' }],
-    };
-    const resultMetadata = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(metadataRequest);
-    expect(resultMetadata.status).toBe(200);
-    expect(resultMetadata.type).toBe('application/json');
-    expect(JSON.parse(resultMetadata.text)).toHaveProperty('metadata');
-    expect(JSON.parse(resultMetadata.text)).toHaveProperty('suggested_fee');
-    expect(JSON.parse(resultMetadata.text).metadata).toHaveProperty('contract_address');
-    expect(JSON.parse(resultMetadata.text).metadata).toHaveProperty('contract_name');
-    expect(JSON.parse(resultMetadata.text).metadata).toHaveProperty('burn_block_height');
-    expect(JSON.parse(resultMetadata.text).suggested_fee[0].value).toBe('260');
-
-    //payloads
-    const contract_address = resultMetadata.body.metadata.contract_address;
-    const contract_name = resultMetadata.body.metadata.contract_name;
-    const burn_block_height = resultMetadata.body.metadata.burn_block_height;
-    const nonce = resultMetadata.body.metadata.account_sequence;
-    const payloadsRequest: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: preprocessRequest.operations, //using operations same as preprocess request
-      metadata: resultMetadata.body.metadata, //using metadata from metadata response
-      public_keys: [
-        {
-          hex_bytes: publicKey,
-          curve_type: 'secp256k1',
-        },
-      ],
-    };
-    const { hashMode, data } = decodeBtcAddress(pox_addr);
-    const hashModeBuffer = bufferCV(new BN(hashMode, 10).toArrayLike(Buffer));
-    const hashbytes = bufferCV(data);
-    const poxAddressCV = tupleCV({
-      hashbytes,
-      version: hashModeBuffer,
-    });
-    const stackingTx: UnsignedContractCallOptions = {
-      contractAddress: contract_address,
-      contractName: contract_name,
-      functionName: 'stack-stx',
-      publicKey: publicKey,
-      functionArgs: [
-        uintCV(stacking_amount),
-        poxAddressCV,
-        uintCV(burn_block_height),
-        uintCV(number_of_cycles),
-      ],
-      validateWithAbi: false,
-      nonce: new BN(nonce), 
-      fee: new BN(fee),
-      network: getStacksNetwork(),
-      anchorMode: AnchorMode.Any,
-    };
-    const transaction = await makeUnsignedContractCall(stackingTx);
-    const unsignedTransaction = transaction.serialize();
-    const signer = new TransactionSigner(transaction);
-    const prehash = makeSigHashPreSign(signer.sigHash, AuthType.Standard, new BN(fee), new BN(nonce));
-    const payloadsResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(payloadsRequest);
-    expect(payloadsResult.status).toBe(200);
-    expect(payloadsResult.type).toBe('application/json');
-    const accountIdentifier: RosettaAccountIdentifier = {
-      address: sender,
-    };
-    const payloadsExpectedResponse = {
-      unsigned_transaction: '0x' + unsignedTransaction.toString('hex'),
-      payloads: [
-        {
-          address: sender,
-          account_identifier: accountIdentifier,
-          hex_bytes: prehash,
-          signature_type: 'ecdsa_recovery',
-        },
-      ],
-    };
-    expect(JSON.parse(payloadsResult.text)).toEqual(payloadsExpectedResponse);
-
-    //combine
-    signer.signOrigin(createStacksPrivateKey(testnetKeys[0].secretKey));
-    const signedSerializedTx = signer.transaction.serialize().toString('hex');
-    const signature: MessageSignature = getSignature(signer.transaction) as MessageSignature;
-    const combineRequest: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      unsigned_transaction: payloadsExpectedResponse.unsigned_transaction,
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes: prehash,
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: publicKey,
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes: signature.data.slice(2) + signature.data.slice(0, 2),
-        },
-      ],
-    };
-    const combineResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(combineRequest);
-    expect(combineResult.status).toBe(200);
-    expect(combineResult.type).toBe('application/json');
-    const combineExpectedResponse: RosettaConstructionCombineResponse = {
-      signed_transaction: '0x' + signedSerializedTx,
-    };
-    expect(combineResult.body).toEqual(combineExpectedResponse);
-
-    //hash
-    const hashRequest: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      signed_transaction: combineExpectedResponse.signed_transaction,
-    };
-
-    const hashResult = await supertest(api.server).post(`/rosetta/v1/construction/hash`).send(hashRequest);
-    expect(hashResult.status).toBe(200);
-
-    const hashExpectedResponse: RosettaConstructionHashResponse = {
-      transaction_identifier: {
-        hash: '0x' + transaction.txid(),
-      },
-    };
-
-    expect(JSON.parse(hashResult.text)).toEqual(hashExpectedResponse);
-
-    //submit
-    const submitRequest: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      // signed transaction bytes
-      signed_transaction: combineResult.body.signed_transaction,
-    };
-    const submitResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/submit`)
-      .send(submitRequest);
-    expect(submitResult.status).toBe(200);
-    expect(submitResult.body.transaction_identifier.hash).toBe(hashExpectedResponse.transaction_identifier.hash);
-
-
-
-    //rosetta block
-    const stxLockedTransaction = await standByForTx(submitResult.body.transaction_identifier.hash);
-
-    const blockHeight = stxLockedTransaction.block_height;
-    let block = await api.datastore.getBlock({ height: blockHeight });
-    assert(block.found);
-    const txs = await api.datastore.getBlockTxsRows(block.result.block_hash);
-    assert(txs.found);
-
-    const blockStxOpsQuery = await supertest(api.address)
-      .post(`/rosetta/v1/block`)
-      .send({
-        network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { hash: block.result.block_hash },
-      });
-    expect(blockStxOpsQuery.status).toBe(200);
-    expect(blockStxOpsQuery.type).toBe('application/json');
-    
-    let stxUnlockHeight = Number.parseInt(blockStxOpsQuery.body.block.transactions[1].operations[1].metadata.unlock_burn_height);
-    expect(stxUnlockHeight).toBeTruthy();
-
-    const blockTxOpsQuery = await supertest(api.address)
-      .post('/rosetta/v1/block/transaction')
-      .send({
-        network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { hash: block.result.block_hash },
-        transaction_identifier: { hash: stxLockedTransaction.tx_id },
-      });
-    expect(blockTxOpsQuery.status).toBe(200);
-    expect(blockTxOpsQuery.type).toBe('application/json');
-
-    const expectedStackStxOp = {
-      type: 'stack_stx',
-      account: {
-        address: testnetKeys[0].stacksAddress,
-      },
-      metadata: {
-        lock_period: number_of_cycles.toString(),
-        amount_ustx: stacking_amount,
-        stacker_address: testnetKeys[0].stacksAddress,
-        pox_addr: pox_addr,
-        start_burn_height: burn_block_height.toString(),
-        unlock_burn_height: stxUnlockHeight.toString(),
-      },
-    };
-
-    const expectedStxLockOp = {
-      type: 'stx_lock',
-      account: {
-        address: testnetKeys[0].stacksAddress,
-      },
-      metadata: {
-        locked: stacking_amount,
-        unlock_height: stxUnlockHeight.toString(),
-      },
-      amount: {
-        value: '-' + stacking_amount,
-        currency: {
-          decimals: 6,
-          symbol: "STX",
-        },
-      },
-    };
-
-    expect(blockStxOpsQuery.body.block.transactions[1].operations).toContainEqual(expect.objectContaining(expectedStackStxOp));
-    expect(blockStxOpsQuery.body.block.transactions[1].operations).toContainEqual(expect.objectContaining(expectedStxLockOp));
-
-    expect(blockTxOpsQuery.body.operations).toContainEqual(expect.objectContaining(expectedStackStxOp));
-    expect(blockTxOpsQuery.body.operations).toContainEqual(expect.objectContaining(expectedStxLockOp));
-
-    let current_burn_block_height = block.result.burn_block_height;
-    //wait for the unlock block height
-    while(current_burn_block_height < stxUnlockHeight){
-      block = await db.getCurrentBlock();
-      assert(block.found);
-      current_burn_block_height =  block.result?.burn_block_height;
-      await timeout(100);
-    }
-
-    const query1 = await supertest(api.address)
-      .post(`/rosetta/v1/block`)
-      .send({
-        network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { hash: block.result.block_hash },
-      });
-
-    expect(query1.status).toBe(200);
-    expect(query1.type).toBe('application/json');
-    expect(query1.body.block.transactions[0].operations[0].type).toBe('coinbase');
-    expect(query1.body.block.transactions[0].operations[1].type).toBe('stx_unlock');
-
-    const query2 = await supertest(api.address)
-    .post(`/rosetta/v1/block/transaction`)
-    .send({
-      network_identifier: { blockchain: 'stacks', network: 'testnet' },
-      block_identifier: { hash: block.result.block_hash },
-      transaction_identifier: {hash: query1.body.block.transactions[0].transaction_identifier.hash }
-    });
-
-    const expectedResponse = {
-      operation_identifier: {
-        index: 1,
-      },
-      type: "stx_unlock",
-      status: "success",
-      account: {
-        address: testnetKeys[0].stacksAddress,
-      },
-      amount: {
-        value: stacking_amount,
-        currency: {
-          decimals: 6,
-          symbol: "STX"
-        }
-      },
-      metadata: {
-        tx_id: query1.body.block.transactions[0].transaction_identifier.hash,
-      }
-    }
-    expect(query2.status).toBe(200);
-    expect(query2.type).toBe('application/json');
-    expect(query2.body.operations[1]).toStrictEqual(expectedResponse);
-
-  })
-
-  test('delegate-stacking rosetta transaction cycle', async() => {
-
-    //derive
-    const publicKey = publicKeyToString(
-      getPublicKey(createStacksPrivateKey(testnetKeys[1].secretKey))
-    );
-    const deriveRequest: RosettaConstructionDeriveRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      public_key: {
-        curve_type: 'secp256k1',
-        hex_bytes: publicKey,
-      },
-    };
-    const deriveResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/derive`)
-      .send(deriveRequest);
-
-    expect(deriveResult.status).toBe(200);
-    expect(deriveResult.type).toBe('application/json');
-
-    const deriveExpectResponse: RosettaConstructionDeriveResponse = {
-      account_identifier: { address: testnetKeys[1].stacksAddress },
-    };
-
-    expect(deriveResult.body).toEqual(deriveExpectResponse);
-
-
-    //preprocess
-    const fee = '260';
-    const stacking_amount = '1250180000000000';//minimum stacking 
-    const sender = deriveResult.body.account_identifier.address;
-    const pox_addr = '2MtzNEqm2D9jcbPJ5mW7Z3AUNwqt3afZH66';
-    const clarityPoxAddr = '0x0c000000020968617368627974657302000000141320e6542e2146ea486700f4091aa793e73607880776657273696f6e020000000101';
-    const size = 253;
-    const max_fee = '12380898';
-    const delegate_to = testnetKeys[2].stacksAddress;
-
-    const preprocessRequest: RosettaConstructionPreprocessRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: [
-
-        {
-          operation_identifier: {
-            index: 0,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'delegate_stx',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: '-'+stacking_amount,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-            metadata: {},
-          },
-          metadata: {
-            pox_addr: pox_addr,
-            delegate_to: delegate_to
-          },
-        },
-        {
-          operation_identifier: {
-            index: 1,
-            network_index: 0,
-          },
-          related_operations: [],
-          type: 'fee',
-          account: {
-            address: sender,
-            metadata: {},
-          },
-          amount: {
-            value: fee,
-            currency: {
-              symbol: 'STX',
-              decimals: 6,
-            },
-          },
-        }
-      ],
-      metadata: {},
-      max_fee: [
-        {
-          value: max_fee,
-          currency: {
-            symbol: 'STX',
-            decimals: 6,
-          },
-          metadata: {},
-        },
-      ],
-      suggested_fee_multiplier: 1,
-    };
-
-    const preprocessResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/preprocess`)
-      .send(preprocessRequest);
-
-    expect(preprocessResult.status).toBe(200);
-    expect(preprocessResult.type).toBe('application/json');
-
-    const expectResponse: RosettaConstructionPreprocessResponse = {
-      options: {
-        fee: fee,
-        sender_address: sender,
-        type: 'delegate_stx',
-        suggested_fee_multiplier: 1,
-        amount: stacking_amount,
-        symbol: 'STX',
-        decimals: 6,
-        max_fee: max_fee,
-        size: size,
-        pox_addr: pox_addr, 
-        delegate_to: testnetKeys[2].stacksAddress
-      },
-      required_public_keys: [
-        {
-          address: sender,
-        },
-      ],
-    };
-    expect(preprocessResult.body).toEqual(expectResponse);
-
-    // //metadata 
-
-    const contract_address = 'ST000000000000000000002AMW42H';
-    const contract_name = 'pox';
-
-    const metadataRequest: RosettaConstructionMetadataRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      options: preprocessResult.body.options, //using options returned from preprocess 
-      public_keys: [{ hex_bytes: publicKey, curve_type: 'secp256k1' }],
-    };
-
-    await standByForPoxToBeReady();
-
-    const resultMetadata = await supertest(api.server)
-      .post(`/rosetta/v1/construction/metadata`)
-      .send(metadataRequest);
-
-      const accountInfo = await new StacksCoreRpcClient().getAccount(sender);
-      const nonce = accountInfo.nonce;
-
-      const metadataResponse: RosettaConstructionMetadataResponse = {
-        metadata: {
-          fee: fee,
-          sender_address: sender,
-          type: 'delegate_stx',
-          suggested_fee_multiplier: 1,
-          amount: stacking_amount,
-          symbol: 'STX',
-          decimals: 6,
-          max_fee: max_fee,
-          delegate_to: testnetKeys[2].stacksAddress,
-          size: size,
-          contract_address: contract_address,
-          contract_name: contract_name,
-          account_sequence: nonce,
-          recent_block_hash: '0x969e494d5aee0166016836f97bbeb3d9473bea8427e477e9de253f78d3212354'
-        },
-        suggested_fee: [ { value: size.toString(), currency: {symbol: 'STX', decimals: 6} } ]
-      }
-      expect(resultMetadata.body).toHaveProperty('metadata');
-      expect(resultMetadata.body.suggested_fee).toStrictEqual(metadataResponse.suggested_fee);
-      expect(resultMetadata.body.metadata.sender_address).toBe(metadataResponse.metadata.sender_address);
-      expect(resultMetadata.body.metadata.type).toBe(metadataResponse.metadata.type);
-      expect(resultMetadata.body.metadata.suggested_fee_multiplier).toBe(metadataResponse.metadata.suggested_fee_multiplier);
-      expect(resultMetadata.body.metadata.amount).toBe(metadataResponse.metadata.amount);
-      expect(resultMetadata.body.metadata.symbol).toBe(metadataResponse.metadata.symbol);
-      expect(resultMetadata.body.metadata.decimals).toBe(metadataResponse.metadata.decimals);
-      expect(resultMetadata.body.metadata.max_fee).toBe(metadataResponse.metadata.max_fee);
-      expect(resultMetadata.body.metadata.delegate_to).toBe(metadataResponse.metadata.delegate_to);
-      expect(resultMetadata.body.metadata.size).toBe(metadataResponse.metadata.size);
-      expect(resultMetadata.body.metadata.contract_address).toBe(metadataResponse.metadata.contract_address);
-      expect(resultMetadata.body.metadata.contract_name).toBe(metadataResponse.metadata.contract_name);
-      expect(resultMetadata.body.metadata.fee).toBe(metadataResponse.metadata.fee);
-      expect(resultMetadata.body.metadata.nonce).toBe(metadataResponse.metadata.nonce);
-      expect(resultMetadata.body.metadata.recent_block_hash).toBeTruthy();//can not predict recent block hash
-
-    //payloads
-    const payloadsRequest: RosettaConstructionPayloadsRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      operations: preprocessRequest.operations, //using operations same as preprocess request
-      metadata: resultMetadata.body.metadata, //using metadata from metadata response
-      public_keys: [
-        {
-          hex_bytes: publicKey,
-          curve_type: 'secp256k1',
-        },
-      ],
-    };
-    const { hashMode, data } = decodeBtcAddress(pox_addr);
-    const hashModeBuffer = bufferCV(new BN(hashMode, 10).toArrayLike(Buffer));
-    const hashbytes = bufferCV(data);
-    const poxAddressCV = tupleCV({
-      hashbytes,
-      version: hashModeBuffer,
-    });
-
-    const stackingTx: UnsignedContractCallOptions = {
-      contractAddress: contract_address,
-      contractName: contract_name,
-      functionName: 'delegate-stx',
-      publicKey: publicKey,
-      functionArgs: [
-        uintCV(stacking_amount),
-        standardPrincipalCV(testnetKeys[2].stacksAddress),
-        noneCV(),
-        someCV(poxAddressCV),
-      ],
-      fee: new BN(fee),
-      nonce: nonce,
-      validateWithAbi: false,
-      network: getStacksNetwork(),
-      anchorMode: AnchorMode.Any,
-    };
-
-    const transaction = await makeUnsignedContractCall(stackingTx);
-    const unsignedTransaction = transaction.serialize();
-    const signer = new TransactionSigner(transaction);
-    const prehash = makeSigHashPreSign(signer.sigHash, AuthType.Standard, new BN(fee), new BN(nonce));
-    const payloadsResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/payloads`)
-      .send(payloadsRequest);
-    expect(payloadsResult.status).toBe(200);
-    expect(payloadsResult.type).toBe('application/json');
-    const accountIdentifier: RosettaAccountIdentifier = {
-      address: sender,
-    };
-    const payloadsExpectedResponse = {
-      unsigned_transaction: '0x' + unsignedTransaction.toString('hex'),
-      payloads: [
-        {
-          address: sender,
-          account_identifier: accountIdentifier,
-          hex_bytes: prehash,
-          signature_type: 'ecdsa_recovery',
-        },
-      ],
-    };
-    expect(JSON.parse(payloadsResult.text)).toEqual(payloadsExpectedResponse);
-
-    //combine
-    signer.signOrigin(createStacksPrivateKey(testnetKeys[1].secretKey));
-    const signedSerializedTx = signer.transaction.serialize().toString('hex');
-    const signature: MessageSignature = getSignature(signer.transaction) as MessageSignature;
-    const combineRequest: RosettaConstructionCombineRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      unsigned_transaction: payloadsExpectedResponse.unsigned_transaction,
-      signatures: [
-        {
-          signing_payload: {
-            hex_bytes: prehash,
-            signature_type: 'ecdsa_recovery',
-          },
-          public_key: {
-            hex_bytes: publicKey,
-            curve_type: 'secp256k1',
-          },
-          signature_type: 'ecdsa_recovery',
-          hex_bytes: signature.data.slice(2) + signature.data.slice(0, 2),
-        },
-      ],
-    };
-    const combineResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/combine`)
-      .send(combineRequest);
-    expect(combineResult.status).toBe(200);
-    expect(combineResult.type).toBe('application/json');
-    const combineExpectedResponse: RosettaConstructionCombineResponse = {
-      signed_transaction: '0x' + signedSerializedTx,
-    };
-    expect(combineResult.body).toEqual(combineExpectedResponse);
-
-    // //hash
-    const hashRequest: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      signed_transaction: combineExpectedResponse.signed_transaction,
-    };
-
-    const hashResult = await supertest(api.server).post(`/rosetta/v1/construction/hash`).send(hashRequest);
-    expect(hashResult.status).toBe(200);
-
-    const hashExpectedResponse: RosettaConstructionHashResponse = {
-      transaction_identifier: {
-        hash: '0x' + transaction.txid(),
-      },
-    };
-
-    expect(JSON.parse(hashResult.text)).toEqual(hashExpectedResponse);
-
-    //submit
-    const submitRequest: RosettaConstructionHashRequest = {
-      network_identifier: {
-        blockchain: RosettaConstants.blockchain,
-        network: getRosettaNetworkName(ChainID.Testnet),
-      },
-      // signed transaction bytes
-      signed_transaction: combineResult.body.signed_transaction,
-    };
-    const submitResult = await supertest(api.server)
-      .post(`/rosetta/v1/construction/submit`)
-      .send(submitRequest);
-
-    expect(submitResult.status).toBe(200);
-    expect(submitResult.body.transaction_identifier.hash).toBe(hashExpectedResponse.transaction_identifier.hash);
-
-
-
-    // //rosetta block
-    const delegateStx = await standByForTx(submitResult.body.transaction_identifier.hash);
-
-    const blockHeight = delegateStx.block_height;
-    let block = await api.datastore.getBlock({ height: blockHeight });
-    assert(block.found);
-    const txs = await api.datastore.getBlockTxsRows(block.result.block_hash);
-    assert(txs.found);
-
-    const blockStxOpsQuery = await supertest(api.address)
-      .post(`/rosetta/v1/block`)
-      .send({
-        network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { hash: block.result.block_hash },
-      });
-    expect(blockStxOpsQuery.status).toBe(200);
-    expect(blockStxOpsQuery.type).toBe('application/json');
-    expect(blockStxOpsQuery.body.block.transactions[1].operations[1]).toMatchObject(
-      {
-        type: 'delegate_stx',
-        account: {
-          address: testnetKeys[1].stacksAddress,
-        },
-        metadata: {
-          amount_ustx: stacking_amount,
-          pox_addr: clarityPoxAddr,
-          delegate_to: delegate_to,
-        },
-      },
-    );
-  })
-
-  test('block/transaction coinbase', async () => {
-    const dbBlock = await db.getCurrentBlock();
-    const blockRes = dbBlock.result ?? undefined;
-    const block = {
-      parent_index_block_hash: blockRes ? blockRes.index_block_hash : '0x00',
-      parent_block_hash: blockRes ? blockRes.block_hash : "0x00",
-      parent_microblock_hash: '',
-      parent_microblock_sequence: 0,
-      block_height: blockRes ? blockRes.block_height + 1 : 1,
-    };
-
-    const tx = {
-      tx_id: '0x1234',
-      tx_index: 4,
-      anchor_mode: AnchorMode.Any,
-      type_id: DbTxTypeId.Coinbase,
-      raw_result: '0x0100000000000000000000000000000001', // u1
-      canonical: true,
-      microblock_canonical: true,
-      microblock_sequence: I32_MAX,
-      microblock_hash: '',
-      parent_index_block_hash: block.parent_microblock_hash,
-      parent_block_hash: block.parent_index_block_hash,
-      post_conditions: Buffer.from([0x01, 0xf5]),
-      fee_rate: 1234n,
-      sender_address: 'sender-addr',
-    };
-
-    const dbMinerReward1 = {
-      canonical: true,
-    };
-
-    const dbMinerReward2 = {
-      canonical: true,
-      recipient: 'testAddr2',
-    };
-
-    const data = new TestBlockBuilder(block)
-      .addTx(tx)
-      .addMinerReward(dbMinerReward1)
-      .addMinerReward(dbMinerReward2)
-      .build();
-
-    await db.update(data);
-    const query1 = await supertest(api.server)
-      .post(`/rosetta/v1/block/transaction`)
-      .send({
-        network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { index: block.block_height, hash: data.block.block_hash },
-        transaction_identifier: { hash: tx.tx_id },
-      });
-    expect(JSON.parse(query1.text)).toEqual({
-      transaction_identifier: {
-        hash: tx.tx_id
-      },
-      operations: [
-        {
-          operation_identifier: {
-            index: 0
-          },
-          type: "coinbase",
-          status: "success",
-          account: {
-            address: tx.sender_address
-          }
-        },
-        {
-          operation_identifier: {
-            index: 1
-          },
-          status: "success",
-          type: "miner_reward",
-          account: {
-            address: dbMinerReward2.recipient,
-          },
-          amount: {
-            value: "21000000000000",
-            currency: {
-              decimals: 6,
-              symbol: "STX"
-            }
-          }
-        },
-        {
-          operation_identifier: {
-            index: 2
-          },
-          status: "success",
-          type: "miner_reward",
-          account: {
-            address: 'testAddr2',
-          },
-          amount: {
-            value: "21000000000000",
-            currency: {
-              decimals: 6,
-              symbol: "STX"
-            }
-          }
-        }
-      ]
-    })
-  });
-
-  /* rosetta construction end */
-
-  afterAll(async () => {
-    await new Promise<void>(resolve => eventServer.close(() => resolve()));
+  afterEach(async () => {
     await api.terminate();
     client.release();
     await db?.close();
